requirements.txt
boto3>=1.28.0
kubernetes>=27.2.0
PyYAML>=6.0
